* Implemented offline-first behavior

+++ app/src/main/java/ua/meugen/android/popularmovies/model/network/req/RateMovieRequest.java
@@ -6,5 +6,5 @@ package ua.meugen.android.popularmovies.model.network.req;
 
 public class RateMovieRequest {
 
-    public double value;
+    public float value;
 }
